requirements.txt
web3
pymongo
python-dotenv
pytesseract
Pillow
transformers
pdfplumber
reportlab
torch